float32 to int

+++ pkg/adt/heap/heap_test.go
@@ -7,7 +7,7 @@ import (
 
 // [29 19 9 4 10 1 2 3 2 2]
 func TestMaxHeap(t *testing.T) {
-	arr := []float32{4, 2, 9, 19, 29, 1, 2, 3, 2, 10}
+	arr := []int{4, 2, 9, 19, 29, 1, 2, 3, 2, 10}
 
 	heap := NewHeap("max", 20)
 
@@ -19,7 +19,7 @@ func TestMaxHeap(t *testing.T) {
 }
 
 func TestMinHeap(t *testing.T) {
-	arr := []float32{4, 2, 9, 19, 29, 1, 2, 3, 2, 10}
+	arr := []int{4, 2, 9, 19, 29, 1, 2, 3, 2, 10}
 
 	heap := NewHeap("min", 20)
